Group authentication pages & remove boilerplate
