+++ src/pages/authentication/Login.jsx
@@ -1,7 +1,8 @@
 import React, { useRef, useState } from 'react'
-import { Form, Button, Card, Container, Alert } from 'react-bootstrap'
-import { useAuth } from '../contexts/AuthContext'
+import { Form, Button, Card, Alert } from 'react-bootstrap'
+import { useAuth } from '../../contexts/AuthContext'
 import { Link } from "react-router-dom"
+import CenteredContainer from '../../components/CenteredContainer'
 
 export function Login() {
   const emailRef = useRef()
@@ -26,8 +27,7 @@ export function Login() {
 
   return (
     <>
-    <Container className="d-flex align-items-center justify-content-center" style={{minHeight: "100vh"}}>
-      <div className="w-100" style={{maxWidth: "400px"}}>
+      <CenteredContainer>
         <Card>
           <Card.Body>
           <h2 className="text-center mb-4">Login</h2>
@@ -48,11 +48,11 @@ export function Login() {
           </div> 
           </Card.Body>
         </Card>
-          <div className="w-100 text-center mt-2">
-            Need an account? <Link to="/register">Sign Up</Link>
-          </div> 
-      </div>
-    </Container>
-  </>
+        <div className="w-100 text-center mt-2">
+          Need an account? <Link to="/register">Sign Up</Link>
+        </div> 
+
+      </CenteredContainer>
+    </>
   )
 }
