+++ src/pages/authentication/PasswordRecovery.jsx
@@ -1,8 +1,9 @@
 import React, { useRef, useState } from 'react'
-import { Alert, Button, Card, Container, Form } from 'react-bootstrap'
+import { Alert, Button, Card, Form } from 'react-bootstrap'
 import { Redirect } from 'react-router'
 import { Link } from "react-router-dom"
-import { useAuth } from '../contexts/AuthContext'
+import CenteredContainer from '../../components/CenteredContainer'
+import { useAuth } from '../../contexts/AuthContext'
 
 export function PasswordRecovery(props) {
   const { confirmResetPassword } = useAuth()
@@ -32,8 +33,7 @@ export function PasswordRecovery(props) {
 
   return (
     <>
-    <Container className="d-flex align-items-center justify-content-center" style={{minHeight: "100vh"}}>
-      <div className="w-100" style={{maxWidth: "400px"}}>
+      <CenteredContainer>
         <Card>
           <Card.Body>
           <h2 className="text-center mb-4">Password Recovery</h2>
@@ -52,8 +52,7 @@ export function PasswordRecovery(props) {
           </Form>
           </Card.Body>
         </Card>
-      </div>
-    </Container>
-  </>
+      </CenteredContainer>
+    </>
   )
 }
