+++ src/pages/authentication/index.jsx
@@ -1,0 +1,4 @@
+export { ForgotPassword } from "./ForgotPassword"
+export { PasswordRecovery } from "./PasswordRecovery"
+export { Login } from "./Login"
+export { Register } from "./Register"
